requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

groq==0.13.0
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.37.1
